SerializableCacheKey
Add SerializableCacheKey and use that in each Key instead of the third party CacheKey. If you put a breakpoint at CompactJavaSerializer.java line 95 and run the first IndicesRequestCache test with -Dtests.security.manager=false, you can see the serialization failing. The IndexReader.CacheKey is in the IndexReader.java class of Lucene.